+++ src/main/java/com/senzing/api/model/SzEntityRecord.java
@@ -1,5 +1,6 @@
 package com.senzing.api.model;
 
+import com.fasterxml.jackson.annotation.JsonInclude;
 import com.senzing.util.JsonUtils;
 
 import javax.json.JsonArray;
@@ -7,6 +8,9 @@ import javax.json.JsonObject;
 import javax.json.JsonString;
 import java.util.*;
 import java.util.function.Consumer;
+
+import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_EMPTY;
+import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 
 /**
  * Describes an entity record.
@@ -28,9 +32,9 @@ public class SzEntityRecord {
   private List<String> addressData;
 
   /**
-   * The list of attribute data strings.
+   * The list of characteristic data strings.
    */
-  private List<String> attributeData;
+  private List<String> characteristicData;
 
   /**
    * The list of identifier data strings.
@@ -69,7 +73,7 @@ public class SzEntityRecord {
     this.dataSource         = null;
     this.recordId           = null;
     this.addressData        = new LinkedList<>();
-    this.attributeData      = new LinkedList<>();
+    this.characteristicData = new LinkedList<>();
     this.identifierData     = new LinkedList<>();
     this.nameData           = new LinkedList<>();
     this.phoneData          = new LinkedList<>();
@@ -123,8 +127,8 @@ public class SzEntityRecord {
    *
    * @return The list of address data strings for the record.
    */
-  public List<String> getAddressData()
-  {
+  @JsonInclude(NON_EMPTY)
+  public List<String> getAddressData() {
     return Collections.unmodifiableList(this.addressData);
   }
 
@@ -150,35 +154,35 @@ public class SzEntityRecord {
   }
 
   /**
-   * Returns the list of attribute data strings for the record.
+   * Returns the list of characteristic data strings for the record.
    *
-   * @return The list of attribute data strings for the record.
+   * @return The list of characteristic data strings for the record.
    */
-  public List<String> getAttributeData()
-  {
-    return Collections.unmodifiableList(this.attributeData);
+  @JsonInclude(NON_EMPTY)
+  public List<String> getCharacteristicData() {
+    return Collections.unmodifiableList(this.characteristicData);
   }
 
   /**
-   * Sets the attribute data list for the record.
+   * Sets the characteristic data list for the record.
    *
-   * @param attributeData The list of attribute data strings.
+   * @param characteristicData The list of characteristic data strings.
    */
-  public void setAttributeData(List<String> attributeData)
+  public void setCharacteristicData(List<String> characteristicData)
   {
-    this.attributeData.clear();
-    this.attributeData.addAll(attributeData);
+    this.characteristicData.clear();
+    this.characteristicData.addAll(characteristicData);
   }
 
   /**
-   * Adds to the attribute data list for the record.
+   * Adds to the characteristic data list for the record.
    *
-   * @param attributeData The attribute data string to add to the address
+   * @param characteristicData The characteristic data string to add to the address
    *                      data list.
    */
-  public void addAttributeData(String attributeData)
+  public void addCharacteristicData(String characteristicData)
   {
-    this.attributeData.add(attributeData);
+    this.characteristicData.add(characteristicData);
   }
 
   /**
@@ -186,8 +190,8 @@ public class SzEntityRecord {
    *
    * @return The list of identifier data strings for the record.
    */
-  public List<String> getIdentifierData()
-  {
+  @JsonInclude(NON_EMPTY)
+  public List<String> getIdentifierData() {
     return Collections.unmodifiableList(this.identifierData);
   }
 
@@ -218,6 +222,7 @@ public class SzEntityRecord {
    *
    * @return The list of name data strings for the record.
    */
+  @JsonInclude(NON_EMPTY)
   public List<String> getNameData() {
     return Collections.unmodifiableList(this.nameData);
   }
@@ -247,6 +252,7 @@ public class SzEntityRecord {
    *
    * @return The list of phone data strings for the record.
    */
+  @JsonInclude(NON_EMPTY)
   public List<String> getPhoneData() {
     return Collections.unmodifiableList(this.phoneData);
   }
@@ -276,6 +282,7 @@ public class SzEntityRecord {
    *
    * @return The list of relationship data strings for the record.
    */
+  @JsonInclude(NON_EMPTY)
   public List<String> getRelationshipData() {
     return Collections.unmodifiableList(this.relationshipData);
   }
@@ -306,6 +313,7 @@ public class SzEntityRecord {
    *
    * @return The list of other data strings for the record.
    */
+  @JsonInclude(NON_EMPTY)
   public List<String> getOtherData() {
     return Collections.unmodifiableList(this.otherData);
   }
@@ -335,6 +343,7 @@ public class SzEntityRecord {
    *
    * @return The original source data that was used to load the record.
    */
+  @JsonInclude(NON_NULL)
   public Object getOriginalSourceData() {
     return originalSourceData;
   }
@@ -448,7 +457,7 @@ public class SzEntityRecord {
     });
 
     getValueList(jsonObject, "ATTRIBUTE_DATA", (attr) -> {
-      rec.addAttributeData(attr);
+      rec.addCharacteristicData(attr);
     });
 
     getValueList(jsonObject, "IDENTIFIER_DATA", (ident) -> {
@@ -483,7 +492,7 @@ public class SzEntityRecord {
     return "dataSource='" + dataSource + '\'' +
         ", recordId='" + recordId + '\'' +
         ", addressData=" + addressData +
-        ", attributeData=" + attributeData +
+        ", characteristicData=" + characteristicData +
         ", identifierData=" + identifierData +
         ", nameData=" + nameData +
         ", phoneData=" + phoneData +