+++ src/main/java/com/senzing/api/model/SzResolutionInfo.java
@@ -1,0 +1,31 @@
+package com.senzing.api.model;
+
+import java.util.List;
+
+public class SzResolutionInfo {
+  /**
+   *
+   */
+  private String dataSource;
+
+  /**
+   *
+   */
+  private String recordId;
+
+  /**
+   *
+   */
+  private List<Long> affectedEntities;
+
+  /**
+   *
+   */
+  private List<SzFlaggedEntity> flaggedEntities;
+
+  /**
+   * Default constructor.
+   */
+  public SzResolutionInfo() {
+  }
+}