+++ src/main/java/com/senzing/api/model/SzWhyRecordsResult.java
@@ -1,11 +1,14 @@
 package com.senzing.api.model;
 
+import com.fasterxml.jackson.annotation.JsonInclude;
 import com.senzing.util.JsonUtils;
 
 import javax.json.JsonArray;
 import javax.json.JsonObject;
 import java.util.ArrayList;
 import java.util.List;
+
+import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 
 /**
  * Describes why an entity resolved.
@@ -44,6 +47,7 @@ public class SzWhyRecordsResult {
    * @return The {@link SzWhyPerspective} identifying and describing the
    *         perspective for this why result from the first record.
    */
+  @JsonInclude(NON_NULL)
   public SzWhyPerspective getPerspective1() {
     return this.perspective1;
   }
@@ -67,6 +71,7 @@ public class SzWhyRecordsResult {
    * @return The {@link SzWhyPerspective} identifying and describing the
    *         perspective for this why result from the first record.
    */
+  @JsonInclude(NON_NULL)
   public SzWhyPerspective getPerspective2() {
     return this.perspective2;
   }
@@ -88,8 +93,9 @@ public class SzWhyRecordsResult {
    *
    * @return The {@link SzMatchInfo} providing the details of the result.
    */
+  @JsonInclude(NON_NULL)
   public SzMatchInfo getMatchInfo() {
-    return matchInfo;
+    return this.matchInfo;
   }
 
   /**